Player movement
The player is now able to move around in both the X and Y axis.

A speed variable will determine how many times the code loops through the change in coords. During each loops the game will check for wall collision to stop the player going off the screen. This will later be used for collision.